Fix project for building multi-target library #1
Ref: xamarin essential
The trick is to set multiple target framework monikers (TFM) in TargetFrameworks. The build system would then be smart enough to build the libraries for all platforms

To make the project referencable via XamarinFormSample, netstandard 2.0 has to be one of the TFM

It is not tested yet as to whether in XamarinFormSample.Android the correct library (i.e. MonoAndroid) would be picked instead of the netstandard one. It should though